Refactor given classes and add new person and teacher class and payable module
