+++ student.rb
@@ -1,16 +1,17 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+class Student < Person
+  # attr_reader :age, :phase
+  # attr_accessor :name
 
   def initialize(options = {})
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+    # @age = options.fetch(:age, 0)
+    # @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
 
   def set_phase(num)
     response = ""
